api/requirements.txt
fastapi
uvicorn
SentenceTransformer

SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.41.1
alembic==1.5.2
psycopg2==2.8.6
psycopg2-binary==2.9.3
pydantic[email]==1.6.1
python-jose[cryptography]==3.2.0
python-multipart==0.0.6
python-dotenv==1.0.0

numpy==1.26.0
jupyter==1.0.0
google-generativeai==0.3.1
